Extract FileTreeFactory from GitFactory
Already needed when constructing a file tree for testing
GitSimple::Utils#glob_to_pathnames, and I think that it is a more
generally clear and useful way to describe the factories.

I have also changed the construction commands to #create and #append
which I think more clearly describe what each step is doing, and can be
applied to different kinds of file trees.