Simplify the method for globing pathnames
I chose to do this because it I was getting a spec failure on the
ordering of the returned values. However, this result was never
actually used so having to fix the spec seemed like a bad idea.

Instead, I have made the following changes:
* return void
* always expect a block to be passed
* change the name from .glob_to_pathname to .each_with_glob, which
  I think will be more clear

Which also avoids needing the spec that was failing locally.