Add automatic garbage collection based on grit
A rugged/libgit2 implementation of garbage collection does not yet
exist, so in order to get it right away I need to make use of the
existing implementation. And since this is already supported by grit I
am not going to re-implement the wrapper around git binary.

grit is unmaintained but I think that it will be OK to use it
temporarily and then replace once I can learn how to implement garbage
collection with libgit2.